refactor: Split Level class into seperate file
